Moving generate tasks to the tasks package
